+++ src/components/Layouts/Header/index.tsx
@@ -1,0 +1,9 @@
+import styles from './styles.module.scss'
+
+export const Header = () => {
+  return (
+    <div className={styles['header-wrapper']}>
+      <header className={styles.header}>header</header>
+    </div>
+  )
+}
