--- src/components/Layouts/Sidebar.tsx
@@ -1,7 +1,0 @@
-export const Sidebar = () => {
-  return (
-    <div>
-      <p>ログイン</p>
-    </div>
-  )
-}
