+++ src/components/Layouts/Sidebar/index.tsx
@@ -1,0 +1,10 @@
+import styles from './styles.module.scss'
+
+export const Sidebar = () => {
+  return (
+    <aside className={styles['side-bar']}>
+      <p>ログイン</p>
+      サイドバー
+    </aside>
+  )
+}
