+++ src/pages/styles.module.scss
@@ -1,0 +1,4 @@
+.contents-wrap {
+  display: flex;
+  justify-content: space-between;
+}
