+++ src/styles/_base.scss
@@ -16,8 +16,8 @@ html:focus-within {
 }
 
 body {
-  background: #fff;
-  color: $c-gray;
+  background: $white;
+  color: $gray;
   font-family: 'Hiragino Sans', sans-serif;
   font-weight: 300;
   line-height: 1.5;
