+++ src/styles/settings/_color.scss
@@ -1,7 +1,40 @@
-$c-bright-apricot: #ffeca6;
-$c-soft-beige: #fff6d3;
-$c-ochre: #b39836;
-$c-bright-orange: #ffc14d;
+/* Setting Layer */
 
-$c-white: #fff;
-$c-gray: #666;
+/* black */
+$black: #000;
+$carbon: #333;
+$black-main: #2d3748;
+$black-dark-grey: #363636;
+
+/* grey */
+$gray: grey;
+$gray-dark: #444;
+$gray-main: #5c616a;
+$gray-light: #999;
+
+/* white */
+$white: #fff;
+$white-main: #f4f4f4;
+$white-light: #fafafa;
+
+/* blue */
+$blue-dark: #1565c0;
+$blue-main: #1976d2;
+$blue-light: #42a5f5;
+$blue-bg: #bfdbfe;
+
+/* green */
+$green-dark: #1b5e20;
+$green-main: #2e7d32;
+$green-light: #4caf50;
+$green-bg: #c6f6d5;
+
+/* red */
+$red-main: #d32f2f;
+$red-light: #ef5350;
+$red-bg: #fecaca;
+
+/* theme */
+$base: #fff6d3;
+$primary: #ffc14d;
+$secondary: #b39836;
