+++ src/styles/settings/_mixin.scss
@@ -1,30 +1,12 @@
-/* Setting Layer */
+$breakpoints: (
+  'sm': 576px,
+  'md': 768px,
+  'lg': 992px,
+  'xl': 1200px,
+);
 
-$iphone5: 320px;
-$sp: 576px;
-$tab: 768px;
-$pc: 1042px;
-
-@mixin iphone5 {
-  @media (max-width: ($iphone5)) {
-    @content;
-  }
-}
-
-@mixin sp {
-  @media (max-width: ($sp)) {
-    @content;
-  }
-}
-
-@mixin tab {
-  @media (max-width: ($tab)) {
-    @content;
-  }
-}
-
-@mixin pc {
-  @media (max-width: ($pc)) {
+@mixin mq($size) {
+  @media screen and (max-width: #{map-get($breakpoints, $size)}) {
     @content;
   }
 }
